refactor: move all to state feature
