--- packages/pluggableWidgets/datagrid-web/src/features/initialization/base.ts
@@ -1,10 +1,0 @@
-import { GridState } from "../../typings/GridState";
-
-export type InitState = GridState;
-
-export interface InitViewState {
-    filter?: undefined;
-    sortOrder?: unknown[];
-}
-
-export type ComputedInitState = [initState: InitState, initView: InitViewState] | [initState: InitState];
