--- packages/pluggableWidgets/datagrid-web/src/features/initialization/init-fresh.ts
@@ -1,7 +1,0 @@
-import { InitState } from "./base";
-import { initGridState } from "../state/grid-state";
-import { GridColumn } from "../../typings/GridColumn";
-
-export function initFresh(props: { columns: GridColumn[] }): [InitState] | undefined {
-    return [initGridState(props.columns)];
-}
