+++ packages/pluggableWidgets/datagrid-web/src/features/state/datasource.ts
@@ -5,8 +5,7 @@ export function hasViewState(ds: ListValue): boolean {
     return ds.sortOrder.length > 0 || ds.filter !== undefined;
 }
 
-export function setViewState(props: { ds: ListValue; initViewState: InitViewState }): void {
-    const { ds, initViewState } = props;
+export function setViewState(ds: ListValue, initViewState: InitViewState): void {
     if (initViewState?.sortOrder) {
         ds.setSortOrder(initViewState.sortOrder as any);
     }
