+++ packages/pluggableWidgets/datagrid-web/src/features/state/init-fresh.ts
@@ -1,0 +1,7 @@
+import { initGridState } from "../state/utils";
+import { GridColumn } from "../../typings/GridColumn";
+import { GridState } from "../../typings/GridState";
+
+export function initFresh(props: { columns: GridColumn[] }): [GridState] | undefined {
+    return [initGridState(props.columns)];
+}
