+++ packages/pluggableWidgets/datagrid-web/src/features/state/init-from-view-state.ts
@@ -1,6 +1,6 @@
 import { ListValue } from "mendix";
 import { ComputedInitState } from "./base";
-import { initGridState } from "../state/grid-state";
+import { initGridState } from "../state/utils";
 import { GridColumn } from "../../typings/GridColumn";
 
 export function initFromViewState(props: { columns: GridColumn[]; ds: ListValue }): ComputedInitState | undefined {
